Rename handler modules to disambiguate from decorator names.
Closes #1162
