+++ starlite/handlers/__init__.py
@@ -1,7 +1,7 @@
-from .asgi import ASGIRouteHandler, asgi
+from .asgi_handlers import ASGIRouteHandler, asgi
 from .base import BaseRouteHandler
-from .http import HTTPRouteHandler, delete, get, head, patch, post, put, route
-from .websocket import WebsocketRouteHandler, websocket
+from .http_handlers import HTTPRouteHandler, delete, get, head, patch, post, put, route
+from .websocket_handlers import WebsocketRouteHandler, websocket
 
 __all__ = (
     "ASGIRouteHandler",
