+++ starlite/routes/websocket.py
@@ -6,7 +6,7 @@ from starlite.routes.base import BaseRoute
 
 if TYPE_CHECKING:
     from starlite.connection import WebSocket
-    from starlite.handlers.websocket import WebsocketRouteHandler
+    from starlite.handlers.websocket_handlers import WebsocketRouteHandler
     from starlite.kwargs import KwargsModel
     from starlite.kwargs.cleanup import DependencyCleanupGroup
     from starlite.types import Receive, Send, WebSocketScope
