+++ tests/connection/websocket/test_websocket.py
@@ -12,7 +12,7 @@ import pytest
 from starlite.connection import WebSocket
 from starlite.datastructures.headers import Headers
 from starlite.exceptions import WebSocketDisconnect, WebSocketException
-from starlite.handlers.websocket import websocket
+from starlite.handlers.websocket_handlers import websocket
 from starlite.status_codes import WS_1001_GOING_AWAY
 from starlite.testing import TestClient, create_test_client
